+++ app/src/main/java/com/example/bodybuilder/data/DailyCalorieData/GainWeightData.kt
@@ -8,5 +8,5 @@ data class GainWeightData(
     @Json(name = "gain weight")
     var gainWeight: String?,
     @Json(name = "calory")
-    var calory: Int?
+    var calory: Float?
 )
